Rewrite config system, fix HTTP poller.
Large rewrite of the config system to fix the HTTP poller so it
inherits top-down configuration successfully.

Added full documentation for the configuration system, example configs
and README.md.

Changed some metric names to remove references to "SSL" and replace with
"TLS", added a "redirect_count" metric to the HTTP poller system.