Add `poller_service_status_boolean` metric
A missing metric in the exporter was an "overall success" metric
similar to what is displayed on the exporter web interface.

`poller_service_status_boolean` returns the value of the internal
Poller.Status() function which is used to populate the web interface,
and represents an overall representation that a service poller, in its
specific configuration, was successful or failed.