tls: Add support for TLS SNI name
TLS SNI was not being sent by default. The `tls_sni_name` option has been
added to allow explicit control and common sense defaults added.

When a hostname is specified and `tls_enabled` is true, the system will
default to sending a TLS SNI name of the hostname.

If a URL is specified with a valid hostname, then the system will send
an SNI name matching the Hostname portion (without the port
specification).

All SNI handling can be overridden by specifying a fixed `tls_sni_name`,
but the default behavior should be unsurprising.